Add bind-utils package because linux.dig
I'm suggesting that the bind-utils package is added on EL7 installs since it's needed for the linux.dig package, which is installed by default. Since it's a default package, it's not clearly noticeable that there is a readme and the linux.dig won't fail until the user tries to run it.